requirements.txt
accelerate==1.6.0
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==4.4.2
diffusers==0.33.1
einops==0.8.1
eva-decord==0.6.1
filelock==3.18.0
fsspec==2025.3.2
hf-xet==1.1.0
huggingface-hub==0.31.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==1.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
proglog==0.1.12
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0